requirements.txt
fastapi~=0.116.1
google-genai~=1.29.0
python-dotenv>=1.0.0
pydantic~=2.11.7
PyYAML~=6.0.2